split js/data into two files & got active/hover state styling works for primary nav links
